Move tetraSpine collisions model into examples
