+++ src/models/obstacles/CMakeLists.txt
@@ -13,6 +13,7 @@ add_library(${PROJECT_NAME} SHARED
         		tgStairs.cpp
 			tgCraterDeep.cpp
 			tgCraterShallow.cpp
+			tgWall.cpp
             )
 
 add_executable(AppObstacleTest
