+++ test_integration/HillTest/CMakeLists.txt
@@ -17,5 +17,5 @@ target_link_libraries(TetraSpineHills_test ${ENV_LIB_DIR}/libgtest.a pthread
 												${NTRT_BUILD_DIR}/examples/IROS_2015/TetraSpineStatic/libtetraSpineHardware.so
                                                 ${NTRT_BUILD_DIR}/examples/IROS_2015/hardwareSineWaves/libtetraSpineLearningSine.so
 												${NTRT_BUILD_DIR}/models/obstacles/libobstacles.so
-												${NTRT_BUILD_DIR}/dev/btietz/tetraCollisions/libtetraCollisions.so)
+												${NTRT_BUILD_DIR}/examples/contactCables/libtetraCollisions.so)
